Add fixed_rating discriminators (#6)
The fixed_rating discriminator family is what enables "dirty branching"
using the Vivaria UI.

This PR also makes updates to the schema. Those were anyway needed in
order to reflect reality, but also in order for Vivaria to accept the
states (e.g. for branching).

One additional major change is that calls to `hooks.save_state` in the
main loop now take in a dict of `{"state": State.dict(), "settings":
Settings.dict()}` instead of simply a `State.dict()`. This is what
Vivaria expects when handling states (e.g. for branching).

Finally, this PR also updates the API name of Claude 3 Haiku and adds
Claude 3.5 Sonnet 20241022 as a claude_legacy_compat model. This is
unrelated to the above.