fix set timeout to handle errors appropriately (#34)
Sorry about the code formatting change that got mixed into this.
The fix is only for the set_timeout function. 

Failed run example: https://mp4-server.koi-moth.ts.net/run/#212714/ss
This failure is caused by set_timeout returning None when the type of
`timeout` is neither a string or an int. We now explicitly handle this
case.